migrated viewstatesrv to ts
